Add rencode module to the deluge library
Deluge uses a variation of Bencode called Rencode for communication with
the Deluge daemon's RPC server. The API and implementation in general is
very rough right now, but it shouldn't matter too much since using this
library won't (or should not) require touching the rencode module at
all. In fact, the rencode module should probably be made private to the
library. It's left as public right now for testing purposes.

As a warning for myself: there are probably a decent number of edge
cases that the code won't handle right now. There is no documentation on
Rencode, so...hopefully nothing breaks too badly.